% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdyns-package.R
\docType{package}
\name{cdyns-package}
\alias{cdyns}
\alias{cdyns-package}
\title{cdyns: Community Dynamics Simulation with Stock Enhancement}
\description{
Community dynamics simulation with stock enhancement. See README for details.
}
\keyword{internal}
