% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_competition}
\alias{set_competition}
\title{Utility: set competition coefficients}
\usage{
set_competition(n_species, int_type, alpha, alpha_scale, inv_sign)
}
\arguments{
\item{n_species}{Number of species in the simulated community.}

\item{int_type}{Method for generating interspecific competition coefficients.
Either \code{"constant"} or \code{"random"}.
\itemize{
\item \code{"constant"}: use a constant value or a supplied full matrix.
\item \code{"random"}: coefficients are drawn from an exponential distribution.
}}

\item{alpha}{Competition coefficient.
\itemize{
\item If \code{int_type = "constant"}, \code{alpha} may be a single value or a full matrix.
\item If \code{int_type = "random"}, \code{alpha} is the expected value of the exponential distribution.
}}

\item{alpha_scale}{Logical; if \code{TRUE}, competition coefficients are scaled by the carrying capacity \code{k}.}

\item{inv_sign}{Logical; if \code{TRUE}, competition coefficients are multiplied by -1.
For instance, in a Ricker model:
\deqn{x \exp(r - \alpha x)}{x \exp(r + \alpha x)}
becomes \deqn{x \exp(r + \alpha x)}{x \exp(r - \alpha x)} if \code{inv_sign = TRUE}.}
}
\description{
Utility: set competition coefficients
}
