% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_im}
\alias{set_im}
\title{Utility: set immigration}
\usage{
set_im(
  n_sim,
  n_species,
  immigration,
  sd_immigration,
  p_immigration,
  stochastic
)
}
\arguments{
\item{n_sim}{Total number of simulation run.}

\item{n_species}{Number of species in the simulated community.}

\item{immigration}{Mean immigration rate per generation.
Immigration follows \eqn{m ~ N(log(immigration), sd_immigration^2)}.}

\item{sd_immigration}{Standard deviation of immigration over time (on a log scale).}

\item{p_immigration}{Probability of immigration occurring in a given time step.}

\item{stochastic}{Logical; if \code{TRUE}, demographic stochasticity is applied by
drawing population and immigration counts from Poisson distributions with
expected means equal to deterministic model outputs.}
}
\description{
Utility: set immigration
}
