% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdynsim.R
\name{cdynsim}
\alias{cdynsim}
\title{Community dynamics simulation with stock enhancement}
\usage{
cdynsim(
  n_timestep = 1000,
  n_warmup = 100,
  n_burnin = 100,
  n_stock_start = NULL,
  n_species = 10,
  k = 100,
  r_type = "constant",
  r = 1.5,
  r_min = 1,
  r_max = 3.5,
  sd_env = 0.1,
  stochastic = FALSE,
  stock = 0,
  phi = 1,
  int_type = "constant",
  alpha = 0.5,
  alpha_scale = "scaled",
  immigration = 0,
  sd_immigration = 0,
  model = "ricker",
  seed = 5,
  seed_interval = 10,
  extinct = 0
)
}
\arguments{
\item{n_timestep}{Number of simulation time steps to be saved}

\item{n_warmup}{Number of warm-up time steps. Species are randomly seeded during this period with no stock enhancement.}

\item{n_burnin}{Number of burn-in time steps. Stock enhancement operates.}

\item{n_stock_start}{Time step at which stocking starts.}

\item{n_species}{Number of species in a simulated community.}

\item{k}{Carrying capacity.}

\item{r_type}{Generation method for intrinsic population growth rates. Either \code{"constant"} or \code{"random"}.}

\item{r}{Intrinsic population growth rate. Disabled if \code{r_type = "random"}.}

\item{r_min}{Minimum value of intrinsic population growth rate. Disabled if \code{r_type = "constant"}.}

\item{r_max}{Maximum value of intrinsic population growth rate. Disabled if \code{r_type = "constant"}.}

\item{sd_env}{SD of environmental stochasticity in a log scale.}

\item{stochastic}{Whether demographic stochasticity is induced or not. If TRUE, population & immigration outcomes will be a random draws from a Poisson distribution with the expected value of population density or immigration.}

\item{stock}{Number of released individuals.}

\item{phi}{Fitness of released individuals relative to wild individuals.}

\item{int_type}{Generation method for an interaction matrix.  Either \code{"constant"}, \code{"random"}, or \code{"manual"}.}

\item{alpha}{Interspecific competition coefficient. Constant if \code{int_type = "constant"}. Expected value of an exponential distribution if \code{int_type = "random"}. Provide a full matrix if \code{int_type = "manual"}.}

\item{immigration}{Mean immigration per generation. Immigration is determined as \code{m ~ N(log(immigration), sd_immigration^2)}}

\item{sd_immigration}{SD immigration over time in a log scale.}

\item{model}{Model for community dynamics. Either \code{"ricker"} (multi-species Ricker model) or \code{"bh"} (multi-species Beverton-Holt model).}

\item{seed}{Expected number of seeds.}

\item{seed_interval}{Time interval for seeding.}

\item{extinct}{Absorbing condition. Species with density < extinct will be removed from the simulation.}
}
\value{
\code{df_dyn}

\code{df_community}

\code{df_species}

\code{interaction_matrix}
}
\description{
Community dynamics simulation with stock enhancement
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
