% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_model}
\alias{fn_model}
\title{Utility: get dynamics function}
\usage{
fn_model(model)
}
\arguments{
\item{model}{Model for community dynamics.
Either \code{"ricker"} (multi-species Ricker model) or
\code{"bh"} (multi-species Beverton-Holt model).}
}
\description{
Utility: get dynamics function
}
