% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_competition}
\alias{set_competition}
\title{Utility: set competition coefficients}
\usage{
set_competition(n_species, int_type, alpha, alpha_scale)
}
\arguments{
\item{n_species}{Number of species in a simulated community.}

\item{int_type}{Generation method for an interaction matrix.  Either \code{"constant"}, \code{"random"}, or \code{"manual"}.}

\item{alpha}{Interspecific competition coefficient. Constant if \code{int_type = "constant"}. Expected value of an exponential distribution if \code{int_type = "random"}. Provide a full matrix if \code{int_type = "manual"}.}

\item{alpha_scale}{"scaled" or "unscaled". If scaled, competition coefficients are scaled by carrying capacity.}
}
\description{
Utility: set competition coefficients
}
